requirements.txt
appdirs==1.4.4
async-generator==1.10
attrs==23.1.0
backports.zoneinfo==0.2.1
beautifulsoup4==4.12.2
blinker==1.6.2
bs4==0.0.1
certifi==2023.5.7
charset-normalizer==3.1.0
click==8.1.3
cssselect==1.2.0
dateparser==1.1.8
demjson3==3.0.6
dnspython==2.3.0
exceptiongroup==1.1.1
facebook-scraper==0.2.59
fake-useragent==1.1.3
Flask==2.3.2
h11==0.14.0
idna==3.4
importlib-metadata==6.6.0
importlib-resources==5.12.0
itsdangerous==2.1.2
Jinja2==3.1.2
lxml==4.9.2
MarkupSafe==2.1.2
outcome==1.2.0
parse==1.19.0
pyee==8.2.2
pymongo==4.3.3
pyppeteer==1.0.2
pyquery==2.0.0
PySocks==1.7.1
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3
regex==2023.5.5
requests==2.30.0
requests-html==0.10.0
selenium==4.9.1
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.4.1
tqdm==4.65.0
trio==0.22.0
trio-websocket==0.10.2
tzlocal==5.0
urllib3==1.26.15
w3lib==2.1.1
websockets==10.4
Werkzeug==2.3.4
wsproto==1.2.0
zipp==3.15.0
